requirements.txt
requests
beautifulsoup4
selenium
Pillow